requirements.txt
streamlit
streamlit-webrtc
streamlit-ace
audio-recorder-streamlit
deepgram-sdk
openai
numpy
opencv-python
pydub
Pillow
transformers
torch
pygame
moviepy
speechrecognition
python-dotenv
langchain-core
phi
groq
google-generativeai
langgraph
langchain
